third_party/blink/web_tests/external/wpt/html/syntax/parsing/html5lib_url_file=tests_innerHTML_1-expected.txt

This is a testharness.js-based test.
Found 1 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] <input><option> (innerHTML in select)
  assert_equals: expected "#document\\n| <option>" but got "#document\\n| <input>\\n| <option>"
Harness: the test ran to completion.
